feat(category-rules): amount-range conditions + fix two critical CRUD bugs (PF-314)
While investigating PF-314's target code, found the existing create and update
endpoints were both completely broken -- neither had ever worked:

- POST /api/v1/category-rules (create): CategoryRuleService.createRule() never
  set an id on the new CategoryRule, so the INSERT query's explicit `:id`
  param bound literal SQL NULL. Explicit NULL bypasses a BIGSERIAL column's
  DEFAULT (only omitting the column entirely lets it apply), so every create
  threw DataIntegrityViolationException on a not-null violation.
- PUT /api/v1/category-rules (update): CategoryRuleRepository never overrode
  JdbcRepository's default update(T), which unconditionally throws
  UnsupportedOperationException. Every update threw on every call.

Neither was caught by the existing suite: CategoryRuleServiceTest is fully
Mockito-mocked (never touches real DB behavior), and the one repository-level
insert test manually set an id, sidestepping the exact broken path.

Fixed both to match the established INSERT/UPDATE convention already used by
AccountRepository/MerchantRepository elsewhere in this codebase:
- INSERT now omits `id` entirely and uses `returning id` + KeyHolder, so
  createRule() returns the real generated id (CategoryRuleController's own
  "id returned" API doc was never actually true before this).
  Repository method renamed insert -> insertAndReturnId since Long doesn't
  satisfy JdbcRepository<T,ID>'s int-returning insert(T) override contract.
- UPDATE is a real, newly-added query, scoped `where id = :id and
  user_id = :userId` -- completing the same three-layer ownership
  defense-in-depth (Controller @PreAuthorize, Service check, SQL-level scope)
  already established for accounts and merchants.

On top of the fixed foundation, added PF-314's actual feature: optional
min_amount/max_amount columns (V37) so a rule can require a transaction's
amount to fall within a range, in addition to its keyword match -- e.g.
"Amazon" under $20 -> Household, over $100 -> Electronics. Evaluated as a
further condition within RuleBasedCategorizationStrategy's existing per-rule
check rather than a separate CategorizationStrategy, since the range only
ever narrows one specific rule's own keyword match. Threaded end-to-end:
migration, domain, DTOs, mapper, row mapper, repository, service, controller,
and the category-rules management UI (form fields + list column) in pf-ui.

Also discovered: src/test/resources/db/test-migration/ is a manually
maintained mirror of db/migration/ (the test Flyway profile points at it
separately, see application.yml) that had already silently drifted -- missing
V34 and V36, neither of which happened to touch a column any test reads or
writes, so nothing caught it. V37 does, which is how this was found. Added
V37 there too; left the pre-existing V34/V36 gap alone as out of scope.

Backend: CategoryRuleRepositoryTest.shouldInsert/shouldUpdate are the real
regression tests proving both bugs are actually fixed (against a live
Testcontainers DB, not mocks) rather than just no-longer-throwing in theory.

+++ src/main/java/com/mayureshpatel/pfdataservice/controller/CategoryRuleController.java
@@ -54,7 +54,7 @@ public class CategoryRuleController {
     @Operation(summary = "Create a category rule", description = "Creates a new category rule for the authenticated user")
     @ApiResponse(responseCode = "200", description = "Rule created, id returned")
     @PostMapping
-    public ResponseEntity<Integer> createRule(
+    public ResponseEntity<Long> createRule(
             @AuthenticationPrincipal CustomUserDetails userDetails,
             @RequestBody @Valid CategoryRuleCreateRequest request) {
         return ResponseEntity.ok(categoryRuleService.createRule(userDetails.getId(), request));
